+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - TableContextConsumer: allows using context inline with the table
 
 ## Changed
+- Always include Search Modal in DOM for accessibility
 
 ## Deprecated